+++ src/main/java/team/unibusk/backend/domain/performance/presentation/PerformanceQueryDocsController.java
@@ -3,25 +3,21 @@ package team.unibusk.backend.domain.performance.presentation;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Encoding;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import team.unibusk.backend.domain.performance.application.dto.response.*;
-import team.unibusk.backend.domain.performance.presentation.request.PerformanceRegisterRequest;
-import team.unibusk.backend.domain.performance.presentation.request.PerformanceUpdateRequest;
 import team.unibusk.backend.global.annotation.MemberId;
-import team.unibusk.backend.global.annotation.SwaggerBody;
 import team.unibusk.backend.global.exception.ExceptionResponse;
 import team.unibusk.backend.global.response.CursorResponse;
 import team.unibusk.backend.global.response.PageResponse;
@@ -31,30 +27,7 @@ import java.util.List;
 
 @Tag(name = "Performance", description = "공연 관련 API")
 @RequestMapping("/performances")
-public interface PerformanceDocsController{
-
-    @Operation(
-            summary = "공연 등록",
-            description = "새로운 공연 정보를 등록합니다. 이미지 파일과 공연 정보를 함께 전송합니다."
-    )
-    @ApiResponses({
-            @ApiResponse(responseCode = "201", description = "공연 등록 성공"),
-            @ApiResponse(responseCode = "400", description = "잘못된 입력 값",
-                    content = @Content(schema = @Schema(implementation = ExceptionResponse.class))),
-            @ApiResponse(responseCode = "401", description = "인증되지 않은 사용자",
-                    content = @Content(schema = @Schema(implementation = ExceptionResponse.class))),
-            @ApiResponse(responseCode = "500", description = "이미지 업로드 실패 등 서버 오류",
-                    content = @Content(schema = @Schema(implementation = ExceptionResponse.class)))
-    })
-    @SwaggerBody(content = @Content(
-            encoding = @Encoding(name = "request", contentType = MediaType.APPLICATION_JSON_VALUE)))
-    @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    ResponseEntity<PerformanceRegisterResponse> registerPerformance(
-            @RequestPart("request") @Valid PerformanceRegisterRequest request,
-            @Parameter(description = "공연 관련 이미지 리스트 (다중 파일 업로드 가능)")
-            @RequestPart(value = "images", required = false) List<MultipartFile> images,
-            @MemberId Long memberId
-    );
+public interface PerformanceQueryDocsController {
 
     @Operation(summary = "다가오는 공연 조회", description = "다가오는 공연을 모두 조회합니다.")
     @ApiResponses({
@@ -101,42 +74,6 @@ public interface PerformanceDocsController{
     })
     @GetMapping("/{performanceId}")
     ResponseEntity<PerformanceDetailResponse> getPerformanceDetail(@PathVariable Long performanceId);
-
-    @Operation(summary = "공연 수정", description = "공연 정보를 수정합니다.")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200", description = "공연 수정 성공"),
-            @ApiResponse(responseCode = "401", description = "인증되지 않은 사용자",
-                    content = @Content(schema = @Schema(implementation = ExceptionResponse.class))),
-            @ApiResponse(responseCode = "403", description = "공연 수정 권한 없음",
-                    content = @Content(schema = @Schema(implementation = ExceptionResponse.class))),
-            @ApiResponse(responseCode = "404", description = "공연을 찾을 수 없음",
-                    content = @Content(schema = @Schema(implementation = ExceptionResponse.class)))
-    })
-    @SwaggerBody(content = @Content(
-            encoding = @Encoding(name = "request", contentType = MediaType.APPLICATION_JSON_VALUE)))
-    @PatchMapping(value = "/{performanceId}", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    ResponseEntity<PerformanceDetailResponse> updatePerformance(
-            @PathVariable Long performanceId,
-            @RequestPart("request") @Valid PerformanceUpdateRequest request,
-            @RequestPart(value = "images", required = false) List<MultipartFile> images,
-            @MemberId Long memberId
-    );
-
-    @Operation(summary = "공연 삭제", description = "공연을 삭제합니다.")
-    @ApiResponses({
-            @ApiResponse(responseCode = "204", description = "공연 삭제 성공"),
-            @ApiResponse(responseCode = "401", description = "인증되지 않은 사용자",
-                    content = @Content(schema = @Schema(implementation = ExceptionResponse.class))),
-            @ApiResponse(responseCode = "403", description = "공연 삭제 권한 없음",
-                    content = @Content(schema = @Schema(implementation = ExceptionResponse.class))),
-            @ApiResponse(responseCode = "404", description = "공연을 찾을 수 없음",
-                    content = @Content(schema = @Schema(implementation = ExceptionResponse.class)))
-    })
-    @DeleteMapping("/{performanceId}")
-    ResponseEntity<Void> deletePerformance(
-            @PathVariable Long performanceId,
-            @MemberId Long memberId
-    );
 
     @Operation(summary = "다가오는 공연 검색", description = "다가오는 공연을 키워드로 검색합니다.")
     @ApiResponses({
